+++ apps/mobile/src/components/watch/Comments.tsx
@@ -3,24 +3,14 @@ import { LENS_CUSTOM_FILTERS } from '@lenstube/constants'
 import type { Publication, PublicationsQueryRequest } from '@lenstube/lens'
 import { useCommentsQuery } from '@lenstube/lens'
 import type { MobileThemeConfig } from '@lenstube/lens/custom-types'
-import { FlashList } from '@shopify/flash-list'
 import { Skeleton } from 'moti/skeleton'
 import type { FC } from 'react'
-import React, { useCallback, useRef } from 'react'
-import {
-  ActivityIndicator,
-  Pressable,
-  StyleSheet,
-  Text,
-  useWindowDimensions,
-  View
-} from 'react-native'
+import React, { useRef } from 'react'
+import { Pressable, StyleSheet, View } from 'react-native'
 
-import normalizeFont from '~/helpers/normalize-font'
 import { useMobileTheme } from '~/hooks'
 
-import NotFound from '../ui/NotFound'
-import Sheet from '../ui/Sheet'
+import CommentsSheet from '../sheets/CommentsSheet'
 import Comment from './Comment'
 import CommentButton from './CommentButton'
 
@@ -38,12 +28,6 @@ const styles = (themeConfig: MobileThemeConfig) =>
       gap: 10,
       height: CONTAINER_HEIGHT,
       width: '100%'
-    },
-    title: {
-      fontFamily: 'font-bold',
-      color: themeConfig.textColor,
-      fontSize: normalizeFont(14),
-      paddingBottom: 10
     }
   })
 
@@ -55,41 +39,19 @@ const Comments: FC<Props> = ({ id }) => {
   const { themeConfig } = useMobileTheme()
   const style = styles(themeConfig)
 
-  const { height } = useWindowDimensions()
   const commentsSheetRef = useRef<BottomSheetModal>(null)
 
   const request: PublicationsQueryRequest = {
-    limit: 10,
+    limit: 1,
     customFilters: LENS_CUSTOM_FILTERS,
     commentsOf: id
   }
 
-  const { data, fetchMore, loading, refetch } = useCommentsQuery({
+  const { data, loading } = useCommentsQuery({
     variables: { request },
     skip: !id
   })
   const comments = data?.publications?.items as Publication[]
-  const pageInfo = data?.publications?.pageInfo
-
-  const fetchMoreComments = async () => {
-    await fetchMore({
-      variables: {
-        request: {
-          ...request,
-          cursor: pageInfo?.next
-        }
-      }
-    })
-  }
-
-  const renderItem = useCallback(
-    ({ item }: { item: Publication }) => (
-      <View style={{ marginTop: 20 }}>
-        <Comment comment={item} richText />
-      </View>
-    ),
-    []
-  )
 
   return (
     <>
@@ -116,34 +78,7 @@ const Comments: FC<Props> = ({ id }) => {
           </Pressable>
         </Skeleton>
       </View>
-
-      <Sheet sheetRef={commentsSheetRef} backdropOpacity={0.9}>
-        <View
-          style={{
-            flex: 1,
-            height: height / 2,
-            paddingHorizontal: 20
-          }}
-        >
-          <Text style={style.title}>Comments</Text>
-
-          <FlashList
-            data={comments ?? []}
-            estimatedItemSize={Boolean(comments?.length) ? comments.length : 5}
-            ListFooterComponent={() =>
-              loading && <ActivityIndicator style={{ paddingVertical: 20 }} />
-            }
-            ListEmptyComponent={() => !loading && <NotFound />}
-            keyExtractor={(item, i) => `${item.id}_${i}`}
-            onEndReachedThreshold={0.8}
-            onEndReached={pageInfo?.next ? fetchMoreComments : null}
-            showsVerticalScrollIndicator={false}
-            renderItem={renderItem}
-            onRefresh={() => refetch()}
-            refreshing={Boolean(comments?.length) && loading}
-          />
-        </View>
-      </Sheet>
+      <CommentsSheet id={id} commentsSheetRef={commentsSheetRef} />
     </>
   )
 }
